+++ lib/Report/Cellular/Step/AggregateSubjectStep.php
@@ -21,22 +21,26 @@ class AggregateSubjectStep extends AggregateRunStep
     {
         $workspace
             ->partition(function (Table $table) {
-                return $table->getAttribute('class') . $table->getAttribute('subject');
+                return $table->getAttribute('identifier');
             })
             ->aggregate(function (Workspace $workspace, $newWorkspace) {
                 if (!$workspace->first()) {
                     return;
                 }
+                $table = $workspace->first();
 
                 if (!isset($newWorkspace[0])) {
                     $newTable = $newWorkspace->createAndAddTable();
+                    foreach (array('class', 'groups', 'subject') as $name) {
+                        $newTable->setAttribute($name, $table->getAttribute($name));
+                    }
                 } else {
                     $newTable = $newWorkspace->first();
                 }
 
-                $table = $workspace->first();
                 $row = $newTable->createAndAddRow();
                 $row->set('iters', $table->count());
+                $row->set('params', $table->getAttribute('parameters'));
                 $row->set('class', $table->getAttribute('class'));
                 $row->set('subject', $table->getAttribute('subject'));
                 $row->set('description', $table->getAttribute('description'));