+++ tests/Functional/Console/Command/RunCommandTest.php
@@ -148,8 +148,9 @@ class RunCommandTest extends BaseCommandTestCase
         $dom = new \DOMDocument();
         $dom->loadXml($display);
         $xpath = new \DOMXPath($dom);
+        $dom->formatOutput = true;
         $benchmarkEls = $xpath->query('//subject');
-        $this->assertEquals(3, $benchmarkEls->length);
+        $this->assertEquals(6, $benchmarkEls->length);
     }
 
     /**