+++ tests/Functional/benchmarks/BenchmarkBench.php
@@ -48,7 +48,8 @@ class BenchmarkBench implements Benchmark
     /**
      * @paramProvider provideParamsOne
      * @paramProvider provideParamsTwo
-     * @description Parameterized bench mark
+     * @description Parameterized bench mark strategy: {strategy}, length: {length}
+     * @group parameterized
      * @iterations 1
      */
     public function benchParameterized(Iteration $iteration)